requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-dotenv
python-multipart
# sqlite3 is part of standard library, no need to list here
